Attempt to fix teams error partial
Still WIP fix, issues still reported

+++ TAC_AI/World/UnloadedBases.cs
@@ -217,7 +217,7 @@ namespace TAC_AI.World
                 DebugTAC_AI.Assert(cand == null, "TileHasEnemy - cand IS NULL");
                 if (cand.tech == null)
                     return false;
-                return AIGlobals.IsBaseTeam(cand.tech.m_TeamID)  
+                return AIGlobals.IsBaseTeamDynamic(cand.tech.m_TeamID)  
                 && ManBaseTeams.IsEnemy(EP.Team, cand.tech.m_TeamID); 
             });
         }
@@ -271,7 +271,8 @@ namespace TAC_AI.World
         {
             if (Singleton.playerTank)
             {
-                if ((tilePos - WorldPosition.FromScenePosition(Singleton.playerTank.boundsCentreWorld).TileCoord).WithinBox(ManEnemyWorld.EnemyRaidProvokeExtents))
+                if ((tilePos - WorldPosition.FromScenePosition(Singleton.playerTank.boundsCentreWorld).TileCoord).
+                    WithinBox(ManEnemyWorld.EnemyRaidProvokeExtents))
                 {
                     return true;
                 }
@@ -281,21 +282,25 @@ namespace TAC_AI.World
 
 
         // Base Operations
+        public static bool CanPurgeTeam(NP_Presence EP, NP_BaseUnit EBU)
+        {
+            return KickStart.CullFarEnemyBases && ManBaseTeams.IsPlayerControlledAIBaseTeam(EP.Team) && 
+                (EBU.tilePos - WorldPosition.FromScenePosition(Singleton.playerPos).TileCoord).WithinBox(
+                    AIGlobals.IgnoreBaseCullingTilesFromOrigin);
+        }
         public static bool PurgeIfNeeded(NP_Presence EP, NP_BaseUnit EBU)
         {
             try
             {
-                if (EBU != null)
+                if (EBU != null && CanPurgeTeam(EP, EBU))
                 {
-                    if (KickStart.CullFarEnemyBases && (EBU.tilePos - WorldPosition.FromScenePosition(Singleton.playerPos).TileCoord).WithinBox(AIGlobals.IgnoreBaseCullingTilesFromOrigin))
+                    // Note: GetBackwardsCompatiblePosition gets the SCENEposition (Position relative to the WorldTreadmillOrigin)!
+                    if (!(EBU.tilePos - WorldPosition.FromScenePosition(Singleton.playerPos).TileCoord).
+                        WithinBox(KickStart.CullFarEnemyBasesDistance))
                     {
-                        // Note: GetBackwardsCompatiblePosition gets the SCENEposition (Position relative to the WorldTreadmillOrigin)!
-                        if (!(EBU.tilePos - WorldPosition.FromScenePosition(Singleton.playerPos).TileCoord).WithinBox(KickStart.CullFarEnemyBasesDistance))
-                        {
-                            DebugTAC_AI.Log("Removing team at " + EBU.tilePos + " since ");
-                            PurgeAllUnder(EP);
-                            return true;
-                        }
+                        DebugTAC_AI.Log("Removing team at " + EBU.tilePos);
+                        PurgeAllUnder(EP);
+                        return true;
                     }
                 }
             }
